In Scala, some classes have more than 2^16 Miranda methods (cherry picked from commit eee27a2b4e91040e27eeb919d174870f307d7aa4)
git-svn-id: https://llvm.org/svn/llvm-project/vmkit/trunk@180457 91177308-0d34-0410-b5e6-96231b3b80d8

+++ lib/j3/VMCore/JavaClass.h
@@ -462,7 +462,7 @@ public:
 
   /// nbVirtualMethods - The number of virtual methods.
   ///
-  uint16 nbVirtualMethods;
+  uint32 nbVirtualMethods;
   
   /// staticMethods - List of all the static methods defined by this class.
   ///
